Suggested edits
Most of the changes are either slightly tweaking the phrasing so that the sentence scans or fixing minor typos.

I'm a little confused why there is both a README.md and docs.md which both cover the same information but not quite.  I think it would be good to just have the README and make sure that it contains all the information required.

Any problems, please let me know.